also allow freeform text in 'seealso' doctags
From the xsl, the purpose of <seealso> tags is to use them with
a 'name' attribute that references another symbol, like:
<seealso name="othersymbol" />

But in the PAWN language guide, in the list of supported XML/HTML
tags, the description of <seealso> is:
"<seealso> </seealso> Text set under the topic “See also”."
which tells it's meant to be freeform text and not have a name
attribute.

This commit ensures that both are accepted: freeform text will
show as paragraphs while occurences that use a "name" attribute
will link to the corresponding symbol.